Added stop function and fixed storage folder creation when sharing media is sent

+++ backend/api/parser/modules/execute.js
@@ -1,25 +1,35 @@
 const { exec } = require('child_process')
 const fs = require('fs')
 
+var playing = false
+
 module.exports = {
     execPresentation: async function (presentationJson) {
+        playing = true
         // if presentation has an audio that will be played the whole time
         if (presentationJson.audiopath != undefined) {
             execAudio(presentationJson.audiopath)
         }
 
         // iterates the slides
+
         for (var i = 0; i < presentationJson.slides.length; i++) {
-            execSlide(presentationJson.slides[i])
-            await sleep(presentationJson.slides[i].duration).then(() => killSlide(presentationJson.slides[i]))
-            //await sleep(3600).then(() => killSlide(presentationJson.slides[i]))
+            if (playing) {
+                execSlide(presentationJson.slides[i])
+                await sleep(presentationJson.slides[i].duration).then(() => killSlide(presentationJson.slides[i]))
+                //await sleep(3600).then(() => killSlide(presentationJson.slides[i]))
 
-            // kills audio if it is the end of the presentation
-            if(i == presentationJson.slides.length -1)
-                exec(`pkill ffplay`)
-
+                // kills audio if it is the end of the presentation
+                if (i == presentationJson.slides.length - 1)
+                    exec(`pkill ffplay`)
+            }
         }
+
     },
+    stop: function () {
+        playing = false
+        exec(`ssh lg${screen.screennumber} "pkill feh pkill mpv pkill ffplay"`)
+    }
 
 }
 
@@ -31,19 +41,19 @@ function killSlide(slide) {
     // kill slide when time is off
     slide.screens.forEach(screen => {
         screen.media.forEach(m => {
-            if(m.type == 'image'){
+            if (m.type == 'image') {
                 //call exec to do ssh and pkill feh
                 exec(`ssh lg${screen.screennumber} "pkill feh"`)
 
-                if(m.sharing != undefined){
+                if (m.sharing != undefined) {
                     exec(`ssh lg${m.partner} "pkill feh"`)
                 }
             }
-            else if(m.type == 'video'){
+            else if (m.type == 'video') {
                 //call exec to do ssh and pkill mpv
                 exec(`ssh lg${screen.screennumber} "pkill mpv"`)
 
-                if(m.sharing != undefined){
+                if (m.sharing != undefined) {
                     exec(`ssh lg${m.partner} "pkill mpv"`)
                 }
             }
@@ -51,7 +61,7 @@ function killSlide(slide) {
     });
 
     // kill slide audio
-    if(slide.audiopath != undefined){
+    if (slide.audiopath != undefined) {
         exec(`pkill ffplay`)
     }
 }
@@ -71,19 +81,19 @@ function execSlide(slide) {
                     if (m.sharing) {
                         openSharedImage(m, screen.screennumber)
                     }
-                    else{
+                    else {
                         openImage(m, screen.screennumber)
                     }
-                    
+
                 }
                 else if (m.type == 'video') {
                     if (m.sharing) {
                         openSharedVideo(m, screen.screennumber)
                     }
-                    else{
+                    else {
                         openVideo(m, screen.screennumber)
                     }
-                    
+
                 }
             }
 
@@ -116,58 +126,58 @@ function openImage(media, screen) {
     else {
         file_path = `${process.env.SLAVE_STORAGE}/${media.storagepath}/${media.filename}`
     }
-    runOpenScript('Image',screen,file_path, media.position)
+    runOpenScript('Image', screen, file_path, media.position)
 }
 
 function openVideo(media, screen) {
     var file_path
-    if(screen == 1){
+    if (screen == 1) {
         file_path = `${process.env.FILE_PATH}/storage/${media.storagepath}/${media.filename}`
     }
     else {
         file_path = `${process.env.SLAVE_STORAGE}/${media.storagepath}/${media.filename}`
     }
 
-    runOpenScript('Video',screen,file_path,media.position)
+    runOpenScript('Video', screen, file_path, media.position)
 }
 
-function runOpenScript(type,screen, file_path, position){
- exec(`${process.env.FILE_PATH}/api/parser/scripts/open${type}.sh ${screen} ${file_path} "${position}"`, (err, stdout, stderr) => {
-    // BUG = EXEC HAS A MAX BUFFER LIMIT, WHEN ACHIEVES IT, STOPS EXECUTION 
-    // putting on background doesn't work, possible solution: increase buffer size 
-    if (err) {
-        //some err occurred
-        console.error(err)
-    } else {
-        // the *entire* stdout and stderr (buffered)
-        console.log(`stdout: ${stdout}`);
-        console.log(`stderr: ${stderr}`);
-    }
-})
+function runOpenScript(type, screen, file_path, position) {
+    exec(`${process.env.FILE_PATH}/api/parser/scripts/open${type}.sh ${screen} ${file_path} "${position}"`, (err, stdout, stderr) => {
+        // BUG = EXEC HAS A MAX BUFFER LIMIT, WHEN ACHIEVES IT, STOPS EXECUTION 
+        // putting on background doesn't work, possible solution: increase buffer size 
+        if (err) {
+            //some err occurred
+            console.error(err)
+        } else {
+            // the *entire* stdout and stderr (buffered)
+            console.log(`stdout: ${stdout}`);
+            console.log(`stderr: ${stderr}`);
+        }
+    })
 }
 
-async function openSharedImage(media,screen) {
-    
-    var leftDest,rightDest
-    
-    if(screen != 1)
+async function openSharedImage(media, screen) {
+
+    var leftDest, rightDest
+
+    if (screen != 1)
         leftDest = `${process.env.SLAVE_STORAGE}/${media.storagepath}/Left${media.filename}`
     else
         leftDest = `${process.env.FILE_PATH}/storage/${media.storagepath}/Left${media.filename}`
 
-    if(media.partner != 1)
+    if (media.partner != 1)
         rightDest = `${process.env.SLAVE_STORAGE}/${media.storagepath}/Right${media.filename}`
     else
         rightDest = `${process.env.FILE_PATH}/storage/${media.storagepath}/Right${media.filename}`
 
-    runOpenScript('Image',screen,leftDest, media.position)
-    runOpenScript('Image',media.partner,rightDest, media.position)
+    runOpenScript('Image', screen, leftDest, media.position)
+    runOpenScript('Image', media.partner, rightDest, media.position)
 
 }
 
 function openSharedVideo(media, screen) {
     console.log('video media/screen', media, screen)
-    
+
 }
 
 function flyTo(destination) {
